fix: media resolving file names
Some groups like Judas upload with AWFUL file names but their torrent names are perfect for resolving. If the media fails we now use the torrent name with the file name to resolve the correct media and episode.

Like... what is "VTuber"? That is not a anime name... "VTuber Legend" or "VTuber Legend: How I went Viral After...." is easier to resolve.