feat: desktop volume boost
Windows, Linux, and MacOS now supports "volume boosting".
If the volume slider in the player is maxed out you will see an icon appear on the right side of the slider that when clicked expands the volume limit.
When you are exceeding the volume limit the slider will be a distinct orange color.