feat: minimize and close miniplayer buttons
The Miniplayer can now be closed using the [X] button, this will unload the torrent and files, but will keep the sidebar button visible now labeled "Last Watched" for easy navigation.
The Miniplayer can now be minimized using the [-] button. This will modify the setting in the Settings > Player Settings > Disable Miniplayer. With the Miniplayer disabled it will be hidden, the icon will change on the sidebar to switch to the Player page instead. The Player will have a new (Eye) icon on the right side of the seekbar that opens the currently playing anime details.
The "Now Playing" icon on the sidebar now has alt-click (or press and hold on Mobile) functionality. This will open the currently playing anime details, so even with the Miniplayer disabled, you can have a shortcut to the anime details without fully opening the Player page.