chore: better descriptive audio
When the Card Audio setting is enabled, the audio displayed on the banner, episode cards, and anime details view will be more descriptive detailing if something is dual audio which is both Dub and Sub, or if its only Sub. Previously only the highest available audio was displayed (Dub if there was a Dub, otherwise it was Sub).