fix: media re-resolving
If the media failed to be resolved then it would re-attempt using the actual torrent name. This is good but it was overwriting media that was successfully resolved.
To correct this, resolved media is mapped and any files that failed to resolve are then searched separately using the torrent name.