fix: video type exclusions
Type exclusions typically would fail to carry over when attempting a second search for files that failed to resolve as the file name would typically be detected as the episode title in this search.
The anime type originally fetched will be remapped to the new results if it doesn't already exist.